Этап 2. Разделение проекта на микросервисы

+++ stats/stats-server/src/main/java/ru/practicum/ewm/mapper/ViewStatsMapper.java
@@ -11,6 +11,6 @@ public class ViewStatsMapper {
         return new ViewStatsDto(
                 viewStats.getApp(),
                 viewStats.getUri(),
-                (int) viewStats.getHits());
+                viewStats.getHits());
     }
 }
